refactor: resolve #1718 — deduplicate ContinuousField trait into fluxion_core/tensor.rs
